move methods to synapse_dynamics_utils
