+++ p8_integration_tests/quick_test/test_stdp/test_IF_cond_exp_stdp_mad_pair_additive.py
@@ -13,11 +13,12 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import spynnaker8 as p
-from p8_integration_tests.base_test_case import BaseTestCase,\
-    calculate_spike_pair_additive_stdp_weight
 import numpy
 import unittest
+import spynnaker8 as p
+from spynnaker.pyNN.models.neuron.synapse_dynamics import (
+    calculate_spike_pair_additive_stdp_weight)
+from p8_integration_tests.base_test_case import BaseTestCase
 
 
 class TestIFCondExpSTDPPairAdditive(BaseTestCase):
