+++ p8_integration_tests/quick_test/test_stdp/test_STDP_pair_additive.py
@@ -14,8 +14,10 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import spynnaker8 as p
-from p8_integration_tests.base_test_case import BaseTestCase,\
-    calculate_spike_pair_additive_stdp_weight
+from spynnaker.pyNN.models.neuron.synapse_dynamics import (
+    calculate_spike_pair_additive_stdp_weight)
+from p8_integration_tests.base_test_case import BaseTestCase
+
 import numpy
