+++ spynnaker/pyNN/models/neuron/synapse_dynamics/__init__.py
@@ -25,11 +25,18 @@ from .synapse_dynamics_stdp import SynapseDynamicsSTDP
 from .synapse_dynamics_structural_common import SynapseDynamicsStructuralCommon
 from .synapse_dynamics_structural_static import SynapseDynamicsStructuralStatic
 from .synapse_dynamics_structural_stdp import SynapseDynamicsStructuralSTDP
+from .synapse_dynamics_utils import (
+    calculate_spike_pair_additive_stdp_weight,
+    calculate_spike_pair_multiplicative_stdp_weight)
+
 
 __all__ = ["AbstractGenerateOnMachine", "AbstractPlasticSynapseDynamics",
            "AbstractStaticSynapseDynamics", "AbstractSynapseDynamics",
-           "AbstractSynapseDynamicsStructural", "PyNNSynapseDynamics",
-           "SynapseDynamicsStatic", "SynapseDynamicsSTDP",
+           "AbstractSynapseDynamicsStructural",
+           "calculate_spike_pair_additive_stdp_weight",
+           "calculate_spike_pair_multiplicative_stdp_weight",
+           "PyNNSynapseDynamics", "SynapseDynamicsStatic",
+           "SynapseDynamicsSTDP",
            # Structural plasticity
            "SynapseDynamicsStructuralCommon",
            "SynapseDynamicsStructuralStatic",
